Register route handlers immediately after their initialization.
Previously the registration of rout handlers (autopage, static content)
happened when app->finish was called; causing the megasplat routes to be
added *after* any other defined route within the app, which allowed for
routes within an app to "mask" the route handlers.

Register the route handlers as part of building the app object so their
routes are always first. Resolves #684.