Remove $app from Dancer2::Logger::File + bugfix
The environment and location attributes are now required and are
provided by the Dancer2::Core::App's _build_logger_engine. This way the
default values of these attributes, which were using $_[0]->app are
removed (as app does not exist inside D2::L::File)

Also the log_dir's trigger method is extracted into it's own. This
allowed the addition of a BUILD method, which calls this extracted
trigger method, which checks the existence of the log_dir (and also
attempts to create it otherwise). Before this fix this check was NOT
executed, when the log_dir was built using _build_log_dir (as the
builder method does not trigger the trigger method).