Add content keyword:
The new 'content' keyword allows to set the content for the
current response. This is to accommodate asynchornous functionality
added in the future.

Until now, setting the content was basically done by returning
content from the route subroutine (using 'template' or data which
is then serialized by a serializer), or you could go to the request
directly - which is generally discouraged, because it goes into
the core without the DSL layer.