Fix redispatching requests losing data:
When using the deprecated dispatcher (using `psgi_app`), requests that
passed between apps (because they haven't matched a route in an App)
will lose their data because the previous matching process already read
the body from the environment.

There are two ways of fixing it:
* Reset the body's filehandle pointer.
* Save the generated request object and use it again.

Because I would be afraid of playing with filehandles and
possibly-cached filehandles, and hell, it would actually make the
dispatching faster because we're caching the request we generated,
and it really is the same request, so there's no reason not to do so.

This fixes GH #723.