Provide all standard responses:
Up until now, for some unbeknownst reason, we had specific methods
for some responses (400, 403, and 404) - only used in core.

While we had Dancer2::Core::HTTP (which uses class methods, *ugh*),
we still didn't use it.

One pull request (GH #733) tried to add methods for all the rest.
Instead we decided to remove the methods and add a single method
that just supports *everything* [that is available in
Dancer2::Core::HTTP].