Bugfix GH#753: seperate content get/set method.
The "before" modifier around the content accessor was called for both
gets and sets. The code was
  * modifying elements in @_, even when they were not variables, causing
    (fatal) warnings about "Modification of a read-only values".
  * when called as a getter, causing $_[1] to become vivified to an undef
    causing the actual accessor to consider it a set.

To solve this we add explicit reader/writer for the content attribute,
and a custom 'content' method to DWIM for get/set calls.

Also corrects the parameterized Maybe[] type syntax.

Resolves most of the "weird" test failures reported in #753.
(leaving the broken tests that involve the serializion of scalars)