Wrap content setting in Response:
Instead of the solution in faaa685158e9432cc2ea3714e26a7df0e72824b7,
we use an "around" modifier to simply wrap the "content" method
to decide whether to serialize or not.

This removes the need for "set_content", "get_content", and having
both an attribute "content" and a method "content" (which itself is
either a setter or getter as well).