Default value for "name" attribute in App class:
Dancer2::Core::App would have to have "name" set for it, but it
wasn't required. So you could ignore it (not provide it) and then
it would fail because it's trying to do stuff with it - which it
*has* to do.

We could require it, but instead we can provide a fair default which
should get the right caller package. It jumps a frame up to get out
of the callback context it is in.