Fix dzil config and TravisCI builds:
We're using the @Dancer Dist::Zilla plugin, which uses the
GithubMeta which added a homepage. We also used the MetaResources
Dist::Zilla plugin which adds the same thing. Builds went nuts.

We're still using @Dancer bundle, but disabling the GithubMeta
one in it because we already indicate everything it provides using
the MetaResources (with the PerlDancer.org website in it).