Use layout_dir attribute, check engine first:
layout_dir was provided after this was written, but merged before,
so it could be used here. By using a configuration, if the user sets
their layout directory somewhere else, it will still work.

We also check the template first to exit early. There's no point to
check it with the template if the template is not available.