Optional prototypes for keywords
Prototypes for keywords will be especially useful for the delayed
response option, being able to change:

    get '/' => sub {
        delayed sub {
            long_action delayed sub {
                content "New";
                done;
            }
        };
    }

to:

    get '/' => sub {
        delayed {
            long_action delayed {
                content "New";
                done;
            }
        };
    }

If a keyword specifies a prototype it is applied via
Scalar::Util::set_prototype (which won't work before perl 5.8.1).

Plugins also import keywords into the *plugin* namespace for D1
backcompat. Ensure the subs (with the deprecation warning) have the
same prototype sig applied as the one they are redefining.