Cover cases fo empty content-length:
I've observed a single case of having a content length which is
an empty string. This failed the test.

Instead we now check for a true content length (not empty strings,
not undefined, and not zero) and then check if it's above zero.

Any fail now will only be a result of having a string for the
content length and we don't want to ignore that.