Do not stringify an expected-overloaded URI object:
Coercions are bad, stringification is bad, and assuming on it is
even worse. Your mileage may vary and doesn't interest me in
this respect. :)

The "content" attribute in the response tried to stringify the
content automatically (using a coerce), which I just removed.

When removing this, a test failed showing that the URI created
by the "uri_for" method returned a blessed string reference with
an overload, which was expected to be stringified.

Fixing this resulted in another test failing. Fixed that test too.