Refresh Serializer::Mutable's format in deserialize(), too
Fixes GH #795.

The 'deserialize' method trusted the format to be already set but
deserialization happens before serialization. So this assumption
was wrong.

The opposite might also be not correct in case there was no request
body and deserialization didn't happen. So the format is refreshed
for both methods.

Also there was kind of a race condition with $serializer->request:
Serializer::Mutable depends on getting the request as an attribute.
But when the Core::Request object is built beforehand its 'BUILD'
method called '$self->data()' for whatever reason. This is fixed by
removing the call and using the lazy character of the 'data' method.

The latter breaks t/deserialize.t.