Fix default content type Serializer::Mutable
Setting the content type as default value for the class
attribute is pointless as the attribute is changed for every
request. Subsequent requests with no matching HTTP headers
will have the last request's content type.

The default for the content type is now set in the
'_get_content_type' method.