Add test for GH #799:
A problem we had was that if you tried to log in core level
with a log format that used the request, the application would
crash because there was no request set at that point yet.

This is a test that verifies it works.

It currently assumes on the Capture logger to also capture the
formatted message, which will be added by a following commit.