Set the request for engines and App earlier:
There is no reason to set the request only when a route is matched,
since it will be the same request no matter which route matched,
and if we have another request cycle, the request will also change.

This caused GH #799 in which, if you use a format that needs a
request, it crashes because the app calls a "core" log in a piece
of code in which the logger engine still hasn't rececived the
request object.