Ensure serializer has request object.
Provides a solution to the chicken-and-egg issue when instantiating a
request object, the serializer needs that request object to deserialize
the body params.

Adding the serializer as part of D2::Core::Request->deserialize feels a
little hacky; but this the simplest approach to solve the immediate need
for Serializer::Mutable.