Remove modification of content_type attribute.
The docs for the content_type attribute in D2::Core::Role::Serializer
state this the the content type *after* serialization. However we were
setting this attribute with (almost) every public method in
Serializer::Mutable, which, well, sucks.

Instead, we give the `content_type` attribute the default
(applicaiton/json), and rename the reader to `default_content_type`
(which seems apt), and add a method `content_type` that checks the
request headers to return the approriate content type we sill serialize
to.

However; there is scope here for some confusion having both an attribute
and method with the same name for anyone working on the code in the
future..