Localize Request and Response:
Now entering crazy-ville.

Instead of the DSL keywords accessing the App's current "request"
and "response" attributes, we're using the route class's variable
identifiers. The DSL now uses those variables to reach the localized
copy.

We also localize the request and response for hooks.

This change will help us remove the attributes altogether in the
future.