Add delayed (asynchronous) responses:
The long-awaited commit is here.

When "compiling" for the user, we change the check for whether
there is a request or not using the $Dancer2::Core::Route::REQUEST
identifier.

We have a new keyword: "delayed", which returns a new response
object (Dancer2::Core::Response::Delayed). Its "to_psgi" method
returns a proper delayed PSGI response subroutine.

The content keyword now will only override the content returned
from a route code if it's simple content and not a reference
of any kind.

We add tests and documentation, because yeah.

We don't support streamed responses yet, but it's close.