Test and eliminate a memory leak with hooks:
When we compile hooks ("compile_hooks" method), we create a closure
that holds on to $self, causing a memory leak.

This wasn't detected in our memory leak test because we didn't
call "to_app" method on it (which finalizes the app and compiles
the hooks).

After calling "to_app", the error shows itself. We then fix it
by closing over a weakened copy.

We also moved $self->request to $Dancer2::Core::Route::REQUEST,
which represents the current localized request.