Handle any level of callback:
The code still didn't support further levels of callbacks. This
means that if you had written:

    use AnyEvent;
    my $timer;
    get '/' => sub {
        delayed sub {
            content 'foo';
            $timer = AE::timer 0.5, 0, sub {
                content 'bar';
            };
        };
    };

The second content (sending "bar") would not have the DSL keywords,
because their localization would be revoked and they would get cleaned
up.

In this commit we're introducing the "delayed" keyword the ability to
detect when it's already in a delayed request context and to instead
wrap a callback with a callback that again localizes closed over variables
so the DSL would work inside the callback.

Now the code can be:

    use AnyEvent;
    my $timer;
    get '/' => sub {
        delayed sub {
            content 'foo';
            $timer = AE::timer 0.5, 0, delayed sub {
                content 'bar';
            };
        };
    };

If you provide a sub instead of a delayed sub, the async will work,
but the DSL keywords would not be available.