Instantiate session before processing template.
Sessions are lazy, that is, no session object is instantiated if the
route does not access the session data. However the template engine is
documented to add the session data to its set of tokens (without extra
futzing from users).

Add a check to test if a session exists but no session object has been
instantiated before processing a template. Ref #840.