Route regexp must have a leading / when prefix is undef.
In general, allowing a route regexp to be constructed by concatenating
the prefixes and route paths gives amazing flexability.
@xsawyerx++ for his work in #817 that allows for this!!

However, there is one edge case when the prefix is undef (or the prefix
is '/' as it is converted to an undef). The PSGI spec required
the PATH_INFO header to be either empty, or begin with a '/'.
Without special handling, a route like
  prefix '/' => sub { get 'x' => sub {...} };
is going to try and match a path of 'x', which is not valid.

Instead, if the prefix is undef and the concatenated regexp does NOT
start with a '/', the route constructor adds it. This implies
  prefix '/' => sub { get '' => sub {...} };
  prefix '/' => sub { get '/' => sub {...} };
both match the path '/', which could cause some confusion, and is a
little different to when the prefix is not undef, viz
  prefix '/foo' => sub { get '' => sub {...} };
  prefix '/foo' => sub { get '/' => sub {...} };
which match '/foo' and '/foo/' respectively.