Error delegates serializing/encoding to response
Let the response object deal with the serialization and encoding of
the error content. By delegating this to the response, the Error
content does not have to concern itself about encoding. The same
applies for serialization, if the response object has a serializer,
it can serialize the Error content.

This commit also moves the default value for the Error `content`
attribute into a builder method as it was getting _very_ long.