Test and eliminate a memory cycle conditionally applying static middleware
The conditionally applied middleware for serving static content was
generating two (concurrent) closures around the psgi app coderef, one
for the conditional, one for the static content middleware itself,
which caused a memory cycle.

The psgi coderef for static content is now constructed separately before
being bassed to the builder of the conditional middleware.