Ensure setting used for public path is consistent
Update Handler::File and its tests to use the setting 'public_dir' from
the apps config for the path to the static content.

Update one last remaining mention of the `public` setting in the manual,
and add a short entry to the migration guide, as D1 uses `public, but d2
uses `public_dir`.