Rework send_file implemenentation
Combine the logic from the existing send_file implementation and
Handler::File, plus add support for passing a filehandle and streaming
support. All in one big commit :)

By using a delayed response, adding streaming support is trivial; The
PSGI env tells us if the server supports it, and the delayed response
allows us to delegate the actual streaming of content to the server.

However, to use  you need to supply a filehandle as the body of the
delayed response. Stole the implementation from Plack::Middleware::Lint
to detect if we are passed a filehandle; of if its a path, we create a
file handle to the static content.

Streaming responses ARE THE DEFAULT if the server supports it. This
should see the memory used by apps using send_file decrease, as they no
longer need to read the content of the file into memory.

You can explicitly pass the option {streaming => 0} to send_file to have
the old behaviour.

Added two new hooks: core.app.before_file_render,
core.app.after_file_render with the same aliases as the old
Handler::File hooks had (they got renamed incase anyone wants to use
Handler::File).

Closes #659