DSL logging keywords call app->log directly.
Avoid a level of indirection my making the DSL logging keywords call
$app->log directly, rather than calling the DSL 'log' keyword.

This is a cleanup to help resolve the "magic number" of call frames to
jump in the logger for reporting what triggered the log message. i.e.
the `log` and `debug`, `info`, `warning` and `error` each need to jump
the same nunmber of call frames to report what triggered the log
message.