Hook execution logs via $app->log
Generate log entries on hook execution by calling $app->log(), rather
than the logging engine directly. This helps minimize the number of
callers involved in generating log messages, which helps minimize the
magic number of call frames to skip when reporting the caller in log
messages.

This approach also provides a single point to deal with future handling of
log messages when no logging engine is defined for an app.