Reduce number of call frames to skip by one
Now the DSL logging keywords call $app->log() directly, we need to skip
one less call frame when reporting the caller.

Updates the concole logger tests, as the 5th stack frame exists - and is
the test file, thanks to Capture::Tiny adding the exact same number of
frames in capturing the output.