GH #879: Remove the requirement for two version numbers:
Dist::Zilla (dzil) was in charge of providing the version number for
the Dancer release, except some test scripts depended on having a
version number too (for example, when testing headers containing the
Server with version).

The issue GH #879 raises is that Dist::Zilla's PkgVersion doesn't
add the version so if dzil's version number is ignored and the one
in the .pm file is used instead.

This commit resolves it by removing the version number everywhere
and any place that requires a version will check for
VERSION || "DUMMY".