Workaround Plugin Hooks problem
As @veryrusty mentioned, plugins are a problem, and there are
changes planned. Nevertheless, I think that the fact of Dancer2
not being able to use two "often used" modules together, like
Dancer2::Plugin::Database and Dancer2::Plugin::Deferred, it also
a big problem.

While we do not do the module rewriting, I am suggesting this
simple workaround. It is not brilliant, it is not efficient, but
it seems to work.