::Role::Hookable requires hook_aliases method
Plugins that add hooks replace the supported_hook and hook_aliases
methods. However these were defined in roles and our role composition
tango makes it unclear which version ended up in the DSL object.

Instead, we require that anything that consumes ::Role::Hookable is
required to define both the supported_hook and hook_aliases methods.

For the DSL, we move those to the ::Core::DSL class, rather than
rely on the composition with ::Role::DSL. This gives predictible
resolution for those two methods, and a workable plugin system.