Add hook_aliases to ::Core::App
Moved from Role::Hookable with extra comments :)

This is the last part of the workaround to get "working" multiple
plugins with hooks. Would have prefered to build to all the hook_aliases
from the core, handlers, engines and plugins.

However, that triggers some of the currently lazy attributes to get
constructed earlier then they currently do. Attempting this broke the
shared_engines tests as the session engine got created using the
defaults. Digging further got messy.. Seems like a reasonable place to
say "plugins behave, but we are still going to rewrite the
implementation".