Catch config loading errors
Config::Any v0.25 returns undef if the files failed to parse; such as
with the invalid YAML in the failure.yml environment test. That undef
can't be used as a hashref. Closes #893.

Config::Any v0.24 and v0.26 throw a parse error on the invalid yaml, so
the assignment never takes place.

Update test to check look for "our" error string, as Config::Any 0.25
does not generate that parse error.