Base Dancer2::Core::Request on Plack::Request:
I couldn't see how to do the majority of this in more than one
commit.

** Why **

The purpose of this commit is to base D2::Core::Request on
Plack::Request instead of delegating to an internal Plack::Request
object.

This allows to utilize the existing implementation of many PSGI
prespectives already available in Plack::Request while adding our
own variations (such as using "is_behind_proxy") without having
to keep another object around and delegating to it.

** How **

Moo was stripped out. Instead we use "parent" to inherit. We
implement all the attributes in very stripped down simple
subroutines, and we copy the "headers_to_array" method from
Dancer2::Core::Role::Headers.

Because of this, "_preq" attribute (which held the Plack::Request
object we delegated to) can be removed, and any method which
called it was simply deleted since it's now available in our
class.

The "BUILDARGS" method allowed multiple ways to instantiate a new
request object, but those were experimental and not very useful,
and basing on "Plack::Request", they added complexity with no
gain. Subsequently, they were removed. Now, in order to create
a new instance, you simply call it with a key/value pair:

    Dancer2::Core::Request->new(
        env => {}, serializer => $s, is_behind_proxy => 0|1
    );

("serializer" and "is_behind_proxy" are not mandatory.)
The documentation reflects this new behavior as well.

All the attributes that are provided by Plack::Request have been
removed. All of the attributes that Plack::Request provides which
we didn't document were documented.

The "forwarded_protocol" method now takes into account all the
headers we know about (which we were checking in the "scheme"
method).

Every method that is incompatible is marked in the code.

The documentation had been almost completely rewritten. All of
the headers are now documented and accounted for, even those
provided by Plack::Request.

There are several additional tasks I've written down that have
to do with additional cleanups we can do.