send_file: streaming logic error & missing import
PSGI streaming is the default option when the server supports it.
However if someone specifiec `streaming => 1` as an option, a logic
error was causing the file *not* to be streamed. Simplified this logic
(outside of the coditional) to make it clearer that streaming *is* the
default.

Also specify the full package name to read_glob_content. Closes #906.