Change references to Dancer to Dancer2 in the DBIC section.
Since this is D2, especially the command line tool reference to
'dancer -a bookstore' should be 'dancer2 -a bookstore' as new users might not
have dancer and would get confused or discouraged. On the other hand,
older users would end up with a Dancer (one) app which might be even worse.