Improve error info when no HTTP_X_FORWARDED_HOST
If behind_proxy is set but no HTTP_X_FORWARDED_HOST header is received
(such as if the proxy's request was from a client using HTTP/1.0), then
Dancer dies with a rather mysterious "Use of uninitialized value in
split" error message. This patch improves the error message to help with
debugging.

Incidentally, this problem is confounded when subsequent requests
are made from the proxy (even with HTTP/1.1) as the environment
variables have already been set. I'm not sure whether this requires a
further patch? If not, this patch at least makes it easier to fix the
problem.