Clean up internal_request when returning from dispatcher
This that die were leaving a request object in the runner when the
dispatcher existed. While this should probably be cleaned up within the
app and or error objects, this ensures we clear the internal_request as
we return from the dispatcher, just in case we've left some dirty state
around.

This should resolve issues noted in #915, and possibly would have
resolved issues with the state of the with_return object from Nov '14.