Allow non-hashref parameters:
Until now parameters had to be a key=value, the way HTTP requires.
However, when using serializers, we might find ourselves receiving
serialized input that is not a hash reference. CBOR is an example
of such a serializer.

Up until now we would crash somewhere internally when trying to
deserialize it. This commit changes it to silently ignoring it,
providing empty parameter values (since there is no "key" to
associate it with), but still making the deserialized data
available in "request->data".

This issue was discussed at GH #740 as part of work to move
Dancer2::Core::Request to be based more on Plack::Request and
adding Hash::MultiValue parameters.

This test uses CBOR::XS and Dancer2::Serializer::CBOR but does
not require them.

This fixes GH #844.