app->set_request propogates to all its engines
There were two places that we here setting the request in the app
object. One of them also set the request in all engine objects, the
other didn't.  The later one then caused template rendering issues when
there was no request object to call params on.

Adds a set_request method that sets the request attribute in the app,
and all its engines when called.

Resolves #936.