Add tests for all session engines.
Tests for the YAML session engine were somehow removed, which resulted
in no tests for ::Role::SessionFactory::File. We now test both the
Simple and the YAML session engine with the same set of tests.

Session files for the YAML engine are written to t/sessions/. This
directory should be empty when we finish - but there is no test for that
at the moment.

I don't like the futzing within the app object that is necessary to replace
the session engine in these tests. Please remember, its done *for
testing only*. Dont to it in your app!