Default session_dir is  appdir/sessions
All the file-based stores are documented that the default directory for
session files is 'appdir/sessions'. However ::SessionFactory::File has a
default of './sessions', which is sane, given its a role that knows
nothing about apps.

When an app instantiates an session engine, pass a default of
session_dir => appdir/sessions, that is still overridable from app
config.

Closes #926.