Refactor: Postpone parameters as much as possible:
The parameters are getting built at least 4 times in a single request.
This is because multiple places call the _build_params() or _params()
(which will build them and cache them).

The only part that actually needs *some* parameters parsed before it
can work is the uploads which needs the body parsed.

What we do instead is call the body parsing from _build_uploads(),
and after that we can call the entire _build_params().

This change will allow us to remove the now-redundant calls to
_build_params().