Optimization: Remove excess _build_params() calls:
After the previous commit's refactor, we can now remove the
unnecessary calls to _build_params().

Only the body parameters are parsed, only when the upload building
occurs, and once all that is done, the rest of them get parsed.

Another possible steps would be to merge only the part that is
necessary, rather than all of them each time.