Optimization: Static middleware is either in or out when creating app:
Until now, each web app was created with a Conditional middleware
that would check, on each request, whether the static handler was
turned on or off.

Instead, we now either add it or not when calling to_app(). There
is no point in adding static handlers *while* running the web
application. Decide before calling to_app().