Optimize: encode content in around modifier, simply code:
The content encoding was done separately from the around method
modifier. This meant that both the around modifier was called
(which had access to the content) and then the encode_content()
method was called, which needed to access the content via the
content() method.

The main thing this commit does is call the encode_content() from
within the "around" method modifier, which already has access to
the content.

Other than that, this commit cleans up the code and reduces the
amount of calls to content().