Use attribute names instead of helper methods:
The engine() helper method validates that the engine access it is
a valid engine that is supported. This is helpful when we read the
configuration and try to apply configurations to a request engine.
We use the engine() method to make sure it's a real one and the
user hasn't made a typo.

However, for internals, since we know exactly which engine we
want to access, there is no need to use the engine() helper method
(which goes through several methods).