\o/ Build engines earlier:
This is finally pushing D2 past D1, and it's a bit of a brain
twister.

Currently the engines are being built lazily and we use predicates
in order to find out whether they exist or not. This is why our
isa checks whether it is a proper object or undef.

All engines (except the serializer) have default implementations
(Template::ForkedTiny, Session::Simple, and Logger::Console)
which means it doesn't make sense for them to have a predicate.

So why do they have a predicate?

The reason is simply that they get built as lazily (== as late)
as possible: during run-time of the first request. This is also
why the performance script in tools/perf.pl runs a single request
first - in order to populate those engines.

However, it doesn't make sense in run-time to change the engines.
If we can assume on the engines not changing once you have created
an application (which makes sense), we can gain quite a bit by
building them at to_app() time. This will allow us to also remove
all the predicated (because they have *some* value - either what
the user has requested or the default).

Well, all of the engines, except the serializer. If we provide a
Null serializer, we could probably simplify the code even further.

What this commit does is:
* Change isa to a simple object, no "maybe".
* Remove predicates for all engines except serializer.
  (Surprisingly, only the logger had a predicate.)
* Add a predicate to the serializer engine.
* Build all engines in to_app().
* defined_engines() is no longer calculated from predicates.
  Instead, we know all engines (except serializer) exist by now,
  and use the serializer predicate to decide whether to add it.
* Use the serializer predicate instead of the attribute where needed.