Build full parameters hash at the end of a match:
And for the probably-final trick: we're going to only build the
parameters hash at the end of a match, since the only code that
gets run and uses the parameters is run *after* a match.

Previously the parameters hash method (_build_params()) was called
when instantiating a new request and then again when we matched
a path to a route. Since it can only be used (even in the "before"
hook) after a match, there's no point in calling it twice.

The operation itself isn't cheap, so this is a big gain.