Better hook and root response content handling
There were two instances where we took a copy of the response content in
another scalar, and passed that to the (now renamed)
_add_content_to_response method, which set it back in the response.

In these cases we can skip the content assignment in the response
(its already there). Setting response content will serialize the content
(if there is a serializer), this approach ensures we do not serialize
response content that has already been serialized.

Resolves #944.