Simply Dancer2::Core::HTTP:
This was some weird code right there. This was one massive hash
with everything in string and numbers and aliases and would
translate between them with awkward code.

Instead, we now have a hash for strings and one for numbers. An
additional change in "behavior" is that if you specify a number
that does not exist, you get undef back, since it's actually a
wrong thing.