GH #947: Tests fail due to unescaped file paths in regex pattern
On Windows, `\` is the directory separator. Therefore, when `t\logger_console.t` is interpolated into a regex pattern, it becomes `tlogger_console.t`,
causing the match, and consequently, the tests in that file to fail.

Even on other platforms, the regex contains two unescaped dots, which, strictly speaking, is not correct.

The solution is to use `\Q$file\E` when interpolating a variable `$file` in to a regex pattern, and use a literal dot in the line number expression.